pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "log_set"
authors = [{name = "Per Rosengren", email = "per.rosengren@gmail.com"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = ["nibe ==0.4.0", "PyYAML ==6.0", "jsonschema ==4.16.0"]

[project.urls]
Source = "https://github.com/per42/nibe_serial"

[project.scripts]
log_set = "log_set:main"
